Removes swappable's requirement for cursorAt
Previously, changing from a sortable to a swappable caused the
unwanted side-effect of changing the cursor and shifting the target
draggable down during the drag. This is because cursorAt was required
by the swappable plugin.  While cursorAt is still an option, the
plugin now by default looks and feels identical to a sortable control,
it just behaves differently.

The technical reason that cursorAt was required was that when not set,
the source element of the drag and the target would be the
same. Represented by itemActive and itemPassive objects in the code,
respectively. The cursorAt setting pushed the cursor outside of the
source element of the drag so on _mouseStop, the event target would be
the element under the cursor (the element to move). Now the
itemPassive object is set to the nearest touching element of the
dragged element, which is the element under the cursor (and underneath
the dragged element).